% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalforecast-package.R
\docType{package}
\name{evalforecast-package}
\alias{evalforecast-package}
\title{evalforecast}
\description{
A package to evaluate forecasters.
}
\author{
Jacob Bien
Maintainer: Jacob Bien<jbien@usc.edu>
}
\keyword{package}
