% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_locations.R
\name{get_forecast_locations_info}
\alias{get_forecast_locations_info}
\title{Get the locations at which to forecast}
\usage{
get_forecast_locations_info(
  df,
  forecast_dates,
  response,
  incidence_period,
  ahead,
  n_locations
)
}
\description{
Get the locations at which to forecast
}
