% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_dates.R
\name{get_or_check_forecast_dates}
\alias{get_or_check_forecast_dates}
\title{If forecast_dates is non-NULL, it makes sure these dates are ok in terms of
backfill buffer.  If NULL, then produces latest possible forecast dates.}
\usage{
get_or_check_forecast_dates(
  df,
  response,
  incidence_period,
  ahead,
  backfill_buffer,
  num_dates = 6,
  forecast_dates = NULL
)
}
\description{
If forecast_dates is non-NULL, it makes sure these dates are ok in terms of
backfill buffer.  If NULL, then produces latest possible forecast dates.
}
