% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_forecasters.R
\name{get_predictions}
\alias{get_predictions}
\title{Return the prediction cards for forecasters that have been run}
\usage{
get_predictions()
}
\value{
a named list of prediction cards such as those generated by [get_prediction_card()]
}
\description{
Return the prediction cards for forecasters that have been run
}
