% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions_card.R
\name{get_predictions_card}
\alias{get_predictions_card}
\title{Get a predictions card}
\usage{
get_predictions_card(
  df,
  your_quantile_forecaster,
  forecast_date,
  forecast_locations = NULL,
  response,
  incidence_period,
  ahead,
  geo_type,
  n_locations
)
}
\arguments{
\item{df}{data frame with columns "location", "reference_date", "issue_date",
"variable_name" and "value".}

\item{your_quantile_forecaster}{a function that follows the template of
my_quantile_forecaster}

\item{forecast_date}{an object of class Date on which forecasts will be made.
e.g. \code{lubridate::ymd("2020-05-04")}}

\item{forecast_locations}{The set of locations where predictions are to be made.
If df contains location_to_be_forecast as a variable_name, then must
leave this argument NULL.  Otherwise, NULL means that predictions at all
locations will be used.}

\item{response}{name of the variable that your_quantile_forecaster is
designed to predict (it should appear in the "variable_name" column of df).
If using incidence_period "epiweek", it should be something for which
summing daily values over an epiweek makes sense (e.g., counts or
proportions but not log(counts) or log(proportions)).}

\item{incidence_period}{one of "epiweek" or "day"}

\item{ahead}{how many epiweeks/days ahead are you forecasting?  If
incidence_period is "epiweek", then ahead = 1 means the epiweek that
includes the forecast date.  If incidence_period is "day" then ahead = 1
means the day after forecast date.}

\item{geo_type}{one of the geographical types, matching the terms used on the
api ("county", "dma", "hrr", "msa", "state")}

\item{n_locations}{How many locations to evaluate predictions at? See
function \code{\link{top_n_locations}} for more details about how these 
locations are chosen.}
}
\value{
a predictions card
}
\description{
Returns a "predictions card" which is a tibble with a row per location and
three columns:
1. location
2. forecast_date
3. forecast_distribution: this is a list column... each element itself
     contains a tibble with the cdc quantiles
A predictions card also has attributes that specify the exact forecasting
task that was being carried out.
}
