% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_response_variables}
\alias{get_response_variables}
\title{Return the correct name of the response variables because of the data change on July 13, 2020}
\usage{
get_response_variables(forecast_date)
}
\arguments{
\item{forecast_date}{the date on which the forecast is being done, such as `lubridate::ymd("2020-07-13")`}
}
\value{
a list of two elements with names `county` and `state` for the respective response variables
}
\description{
Return the correct name of the response variables because of the data change on July 13, 2020
}
