% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_locations.R
\name{get_top_n_locations}
\alias{get_top_n_locations}
\title{Get the top n (or more) locations}
\usage{
get_top_n_locations(df, response, latest_date, n)
}
\arguments{
\item{df}{data frame with columns location, reference_date, issue_date,
variable_name, and value (and possibly others, which we will ignore)}

\item{response}{string indicating the name of the response}

\item{latest_date}{the latest date which we will use when making our location
choice}

\item{n}{how many locations should we pick? If it is greater than the total
number of locations, than we pick all of them.}
}
\description{
This function returns the n** locations which have observed the largest
values of the response, cumulatively over all dates up to and including the
latest date
}
\details{
**: (if there are multiple locations with value equal to the nth location,
they are all included)
}
