% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_cards.R
\name{locate_cards}
\alias{locate_cards}
\title{Locate all the predictions cards or scorecards in a directory}
\usage{
locate_cards(
  cards_dir,
  dir_order = c("forecast_date", "ahead", "response", "geo_type", "incidence_period",
    "n_locations", "forecaster"),
  cards_type = c("scorecard.RDS", "out.RDS")
)
}
\arguments{
\item{cards_dir}{name of directory with the structure described above}

\item{dir_order}{specifies the order of the directories}

\item{cards_type}{either "scorecard.RDS", for scorecards or "out.RDS", for 
predictions cards.}
}
\description{
Given a structured directory tree that contains predictions cards or
scorecards saved as RDS files, this function prepares a convenient data frame
with the information contained in the directory tree structure and locates
the file names.  The default directory structure is
}
\details{
forecast_date/ahead/response/geo_type/incidence_period/n_locations/forecaster/scorecard.RDS

although this can be changed using the 
second argument.
}
