% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_forecaster.R
\name{run_forecaster}
\alias{run_forecaster}
\title{Test a forecaster if it works using the prediction harness}
\usage{
run_forecaster(
  response,
  incidence_period = c("epiweek"),
  ahead,
  forecast_date,
  geo_type = c("county", "state", "hrr", "msa"),
  n_locations = 200,
  data_path,
  forecaster
)
}
\arguments{
\item{response}{the response (e.g. "usafacts_deaths_incidence_num")}

\item{incidence_period}{the incidence period (e.g. "epiweek" for
now, for all forecasters)}

\item{ahead}{the ahead parameter (e.g. 1, 2, 3, 4)}

\item{forecast_date}{the date of the forecast}

\item{geo_type}{the geographic type (e.g "county" or "state" or
"hrr" or "msa"... but for now only the first two),}

\item{n_locations}{the number of locations (for now we will use 200
for this)}

\item{data_path}{full path to your county/state Rdata file}

\item{forecaster}{your forecaster function}
}
\description{
Test a forecaster if it works using the prediction harness
}
