% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_forecasters.R
\name{run_forecasters}
\alias{run_forecasters}
\title{Run forecasters in specified list and populate a global variable with prediction_cards incrementally}
\usage{
run_forecasters(
  forecasters,
  response,
  incidence_period = c("epiweek"),
  ahead,
  forecast_date,
  geo_type = c("county", "state", "hrr", "msa"),
  n_locations = 200,
  data_list
)
}
\arguments{
\item{forecasters}{a named list of forecasters, such as those returned by [get_forecaster()]}

\item{response}{the response (e.g. "usafacts_deaths_incidence_num")}

\item{incidence_period}{the incidence period (e.g. "epiweek" for
now, for all forecasters)}

\item{ahead}{the ahead parameter (e.g. 1, 2, 3, 4)}

\item{forecast_date}{the date of the forecast}

\item{geo_type}{the geographic type (e.g "county" or "state" or
"hrr" or "msa"... but for now only the first two),}

\item{n_locations}{the number of locations (for now we will use 200
for this)}

\item{data_list}{a named list of data frames for each response/geo_type, i.e. for county/state}
}
\description{
Run forecasters in specified list and populate a global variable with prediction_cards incrementally
}
