% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_production.R
\name{run_production}
\alias{run_production}
\title{Run all forecasters using the production data}
\usage{
run_production(
  forecaster_pkgs,
  forecaster_filter = list(type = function(type) type == "standalone", name =
    function(name) TRUE),
  output_path,
  county_data_path,
  state_data_path,
  forecast_date,
  git_sha,
  aheads = 1:4,
  geo_types = c("county", "state"),
  responses,
  incidence_period = "epiweek",
  n_locations = 200
)
}
\arguments{
\item{forecaster_pkgs}{a string array of forecasting R package names}

\item{output_path}{the output path where all output should be stored in a hierarchy}

\item{county_data_path}{the full path of the county data file}

\item{state_data_path}{the full path of the state data file}

\item{forecast_date}{the forecast data (yyyy-mm-dd)}

\item{git_sha}{the sha for the github repo from which packages where installed}

\item{responses}{a named list of responses for county and state}

\item{incidence_period}{the incidence period, default "epiweek"}

\item{n_locations}{the number of locations to forecast for, default 200}

\item{forecaster_filters}{a list of conjunctive criteria (implemented as functions of `type` and `name`) to filter the forecasters}

\item{ahead}{the value of ahead to use, default `1:4`}

\item{geo_type}{the geo_type values to run for, default `c("county", "state")`}
}
\description{
Run all forecasters using the production data
}
